Refactor api code
