--- fc_app/__init__.py
@@ -1,1 +1,0 @@
-from .server import FCServer
